Version 0.0.6
- Fix compiler error for ambiguous requestFrom with 1.5.0-rc.2.

+++ src/SC16IS740RK.cpp
@@ -235,7 +235,7 @@ bool SC16IS740::readInternal(uint8_t *buffer, size_t size) {
 	wire.write(RHR_THR_REG << 3);
 	wire.endTransmission(false);
 
-	uint8_t numRcvd = wire.requestFrom(addr, size, true);
+	uint8_t numRcvd = wire.requestFrom(addr, size, (uint8_t)true);
 	if (numRcvd < size) {
 		log.info("readInternal failed numRcvd=%u size=%u", numRcvd, size);
 		return false;
